[Xtensa] Fix Clang builtins include directory
This fixes https://github.com/espressif/llvm-project/issues/83

In short, it adjusts the include path logic to include the builtins
directory as long as `-nostdinc` or `-nobuiltininc` isn't specified.
Previously, the builtins directory would not be included if either the
GCC installation wasn't found, or `-nostdlibinc` was specified (both of
which aren't related to the builtins directory).